Script error when loading saves
m_UnitId causes a script error when loading a save if there has been
changes to what units are in base army view.
Saving/loading appears to be working flawlessly even with the two lines
related to m_UnitId commented out so if anyone has ideas on why it was
included in the first place I would be interested in hearing them.